requirements.txt
asgiref==3.3.1
astroid==1.6.5
autopep8==1.3.5
boto3==1.17.49
botocore==1.20.49
certifi==2018.4.16
chardet==3.0.4
charset-normalizer==2.0.11
colorama==0.4.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.2.17
django-admin-rangefilter==0.3.8
django-data-exports==0.7
django-excel==0.0.10
django-exportdata==0.2
django-filebrowser-no-grappelli==3.7.4
django-heroku==0.3.1
django-import-export==1.1.0
django-inspect-model==0.7
django-ipware==3.0.2
django-model-changes==0.15
django-rest-multiple-models==2.1.3
django-s3direct==1.1.8
django-sendsms==0.4
django-social-share==2.2.1
django-storages==1.6.6
django-tinymce4-lite==1.7.0
django-user-agents==0.3.2
DjangoModelTracker==1.3.9
djangorestframework==3.8.2
docutils==0.14
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.7
gcloud==0.18.3
googleapis-common-protos==1.5.3
gunicorn==20.0.4
httplib2==0.11.3
idna==2.6
isort==4.3.4
jdcal==1.4
jmespath==0.9.3
jsmin==3.0.1
jsonfield==3.1.0
jws==0.1.3
lazy-object-proxy==1.3.1
lml==0.1.0
maxminddb==2.0.3
mccabe==0.6.1
Naked==0.1.31
oauth2client==4.1.3
odfpy==1.3.6
openpyxl==2.5.9
pep8==1.7.1
Pillow==9.0.0
protobuf==3.5.2.post1
psycopg2==2.8.6
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycodestyle==2.5.0
pycryptodome==3.14.0
pycryptodomex==3.14.0
pyexcel==0.6.6
pyexcel-io==0.6.4
pyexcel-webio==0.1.4
pyflakes==2.1.1
pylint==1.9.2
Pyrebase4==4.5.0
python-dateutil==2.7.3
python-env-utils==0.4.1
python-gettext==3.0
python-jwt==2.0.1
pytz==2018.3
PyYAML==5.4.1
requests==2.27.1
requests-toolbelt==0.9.1
rsa==3.4.2
s3transfer==0.3.6
shellescape==3.8.1
simplejson==3.17.2
six==1.11.0
sqlparse==0.4.1
tablib==0.12.1
texttable==1.6.3
ua-parser==0.8.0
unicodecsv==0.14.1
urllib3==1.26.3
user-agents==2.0
whitenoise==5.2.0
wrapt==1.10.11
xlrd==1.1.0
xlwt==1.3.0
